move source files to src file, add doc strings to functions, classes, and methods.
